# ResumeAPI/Schemas/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ResumeAPI",
    "version": "v1"
  },
  "paths": {
    "/resume/build/{id}": {
      "get": {
        "tags": [
          "Resume"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/resume/build": {
      "post": {
        "tags": [
          "Resume"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Resume"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Resume"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Resume"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/resume/build-test/{id}": {
      "get": {
        "tags": [
          "Resume"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/resume/build-test": {
      "get": {
        "tags": [
          "Resume"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/resume/get-all": {
      "get": {
        "tags": [
          "Resume"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResumeTreeNode"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResumeTreeNode"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResumeTreeNode"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/resume/get/{id}": {
      "get": {
        "tags": [
          "Resume"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeTreeNode"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeTreeNode"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeTreeNode"
                }
              }
            }
          }
        }
      }
    },
    "/resume/upsert": {
      "post": {
        "tags": [
          "Resume"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeTreeNode"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeTreeNode"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeTreeNode"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeTreeNode"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeTreeNode"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeTreeNode"
                }
              }
            }
          }
        }
      }
    },
    "/resume/delete/{id}": {
      "delete": {
        "tags": [
          "Resume"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/users/create": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Create user",
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "Delete User",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "UserId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/healthcheck": {
      "get": {
        "tags": [
          "ApiHealth"
        ],
        "description": "Health check endpoint",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "PhoneNumber": {
        "type": "object",
        "properties": {
          "areaCode": {
            "type": "integer",
            "format": "int32"
          },
          "prefix": {
            "type": "integer",
            "format": "int32"
          },
          "lineNumber": {
            "type": "integer",
            "format": "int32"
          },
          "formattedNumber": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "Resume": {
        "type": "object",
        "properties": {
          "header": {
            "$ref": "#/components/schemas/ResumeHeader"
          },
          "education": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResumeEducation"
            },
            "nullable": true
          },
          "experience": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResumeExperience"
            },
            "nullable": true
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "splitResume": {
            "type": "boolean"
          },
          "splitExperienceAfter": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ResumeDegree": {
        "type": "object",
        "properties": {
          "typeOfDegree": {
            "type": "string",
            "nullable": true
          },
          "major": {
            "type": "string",
            "nullable": true
          },
          "minor": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ResumeEducation": {
        "type": "object",
        "properties": {
          "school": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "graduationYear": {
            "type": "string",
            "nullable": true
          },
          "degree": {
            "$ref": "#/components/schemas/ResumeDegree"
          }
        },
        "additionalProperties": false
      },
      "ResumeExperience": {
        "type": "object",
        "properties": {
          "jobTitle": {
            "type": "string",
            "nullable": true
          },
          "employer": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "responsibilities": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ResumeHeader": {
        "type": "object",
        "properties": {
          "filename": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "website": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "$ref": "#/components/schemas/PhoneNumber"
          },
          "summary": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ResumeTreeNode": {
        "type": "object",
        "properties": {
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResumeTreeNode"
            },
            "nullable": true
          },
          "comments": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "active": {
            "type": "boolean"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "parentId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "nodeType": {
            "type": "string",
            "nullable": true
          },
          "depth": {
            "type": "integer",
            "format": "int32"
          },
          "order": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      }
    }
  }
}